refactor: 将 labels 移至 geometry/label 下
